Misc cleanups
