+++ src/Typhon.Engine/Data/Index/BTree.LookupStatus.cs
