+++ src/Typhon.Engine/Data/Index/BTree.cs
@@ -4,6 +4,7 @@ using JetBrains.Annotations;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Threading;
@@ -18,7 +19,7 @@ using System.Threading;
 //
 ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-namespace Typhon.Engine.BPTree;
+namespace Typhon.Engine;
 
 #region Chunk definitions
 
@@ -162,10 +163,7 @@ public abstract partial class BTree<TKey> : IBTree where TKey : unmanaged
         public TKey Key;
         public int Value;
 
-        public static void ChangeKey(ref KeyValueItem item, TKey newKey)
-        {
-            item = new KeyValueItem(newKey, item.Value);
-        }
+        public static void ChangeKey(ref KeyValueItem item, TKey newKey) => item = new KeyValueItem(newKey, item.Value);
 
         public static void SwapKeys(ref KeyValueItem x, ref KeyValueItem y)
         {
@@ -183,7 +181,7 @@ public abstract partial class BTree<TKey> : IBTree where TKey : unmanaged
             _keyComparer = comparer ?? Comparer<TKey>.Default;
             Key = key;
             Added = false;
-            ElementId = default;
+            ElementId = 0;
             Accessor = ref accessor;
         }
         public readonly TKey Key;
@@ -253,7 +251,7 @@ public abstract partial class BTree<TKey> : IBTree where TKey : unmanaged
             Key = key;
             Comparer = comparer;
 
-            Value = default;
+            Value = 0;
             Removed = false;
             Accessor = ref accessor;
         }
@@ -724,13 +722,13 @@ public abstract partial class BTree<TKey> : IBTree where TKey : unmanaged
     private NodeWrapper Root
     {
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        get => _rootChunkId == 0 ? default : new NodeWrapper(_storage, _rootChunkId);
+        get => _rootChunkId == 0 ? default : new NodeWrapper(_storage, _rootChunkId, _height == 1);
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         set => _rootChunkId = value.IsValid ? value.ChunkId : 0;
     }
 
-    private NodeWrapper LinkList;
-    private NodeWrapper ReverseLinkList;
+    private NodeWrapper _linkList;
+    private NodeWrapper _reverseLinkList;
 
     // Volatile height: atomically readable by concurrent readers under OLC.
     // Only modified under exclusive lock; volatile prevents compiler reordering for readers.
@@ -744,8 +742,8 @@ public abstract partial class BTree<TKey> : IBTree where TKey : unmanaged
         set => _height = value;
     }
 
-    protected KeyValueItem GetFirst(ref ChunkAccessor accessor) => LinkList.GetFirst(ref accessor);
-    protected KeyValueItem GetLast(ref ChunkAccessor accessor) => ReverseLinkList.GetLast(ref accessor);
+    protected KeyValueItem GetFirst(ref ChunkAccessor accessor) => _linkList.GetFirst(ref accessor);
+    protected KeyValueItem GetLast(ref ChunkAccessor accessor) => _reverseLinkList.GetLast(ref accessor);
 
     #endregion
 
@@ -809,7 +807,7 @@ public abstract partial class BTree<TKey> : IBTree where TKey : unmanaged
                         node = node.GetLeft(ref accessor);
                         Height++;
                     }
-                    LinkList = node;
+                    _linkList = node;
 
                     // Traverse the rightmost path to find ReverseLinkList (rightmost leaf)
                     node = Root;
@@ -817,7 +815,7 @@ public abstract partial class BTree<TKey> : IBTree where TKey : unmanaged
                     {
                         node = node.GetLastChild(ref accessor);
                     }
-                    ReverseLinkList = node;
+                    _reverseLinkList = node;
                 }
             }
             finally
@@ -861,7 +859,7 @@ public abstract partial class BTree<TKey> : IBTree where TKey : unmanaged
     /// </summary>
     private unsafe void FindInDirectory(short stableId, ref ChunkAccessor accessor)
     {
-        var chunk0Addr = accessor.GetChunkAddress(0, false);
+        var chunk0Addr = accessor.GetChunkAddress(0);
         ref var header = ref Unsafe.AsRef<BTreeDirectoryHeader>(chunk0Addr);
 
         int totalEntries = header.EntryCount;
@@ -870,7 +868,7 @@ public abstract partial class BTree<TKey> : IBTree where TKey : unmanaged
         for (var i = 0; i < totalEntries; i++)
         {
             var (chunkId, offset) = ComputeEntryLocation(i, stride);
-            var entryChunkAddr = accessor.GetChunkAddress(chunkId, false);
+            var entryChunkAddr = accessor.GetChunkAddress(chunkId);
             ref var entry = ref Unsafe.AsRef<BTreeDirectoryEntry>(entryChunkAddr + offset);
 
             if (entry.StableId == stableId)
@@ -992,12 +990,12 @@ public abstract partial class BTree<TKey> : IBTree where TKey : unmanaged
 
         // Check the linked link of leaves in forward
         NodeWrapper prev = default;
-        var cur = LinkList;
+        var cur = _linkList;
         TKey prevValue = default;
 
         while (cur.IsValid)
         {
-            if (cur != LinkList)
+            if (cur != _linkList)
             {
                 ConsistencyAssert(prev.GetNext(ref accessor) == cur, "Prev.Next doesn't link to current");
                 ConsistencyAssert(cur.GetPrevious(ref accessor) == prev, "Cur.Previous doesn't link to previous");
@@ -1010,16 +1008,16 @@ public abstract partial class BTree<TKey> : IBTree where TKey : unmanaged
             prev = cur;
             cur = cur.GetNext(ref accessor);
         }
-        ConsistencyAssert(prev == ReverseLinkList, "Last Node of the forward chain doesn't match ReverseLinkList");
+        ConsistencyAssert(prev == _reverseLinkList, "Last Node of the forward chain doesn't match ReverseLinkList");
 
         // Check the linked link of leaves in reverse
         NodeWrapper next = default;
-        cur = ReverseLinkList;
+        cur = _reverseLinkList;
         TKey nextValue = default;
 
         while (cur.IsValid)
         {
-            if (cur != ReverseLinkList)
+            if (cur != _reverseLinkList)
             {
                 ConsistencyAssert(next.GetPrevious(ref accessor) == cur, "Next.Previous doesn't link to current");
                 ConsistencyAssert(cur.GetNext(ref accessor) == next, "Cur.Next doesn't link to next");
@@ -1032,9 +1030,10 @@ public abstract partial class BTree<TKey> : IBTree where TKey : unmanaged
             next = cur;
             cur = cur.GetPrevious(ref accessor);
         }
-        ConsistencyAssert(next == LinkList, "Last Node of the reverse chain doesn't match LinkedList");
+        ConsistencyAssert(next == _linkList, "Last Node of the reverse chain doesn't match LinkedList");
     }
 
+    [ExcludeFromCodeCoverage]
     private static void ConsistencyAssert(bool condition, string message)
     {
         if (!condition)
@@ -1306,34 +1305,17 @@ public abstract partial class BTree<TKey> : IBTree where TKey : unmanaged
         return node;
     }
 
-    /// <summary>Result of an OLC insert attempt.</summary>
-    private enum OlcInsertResult
-    {
-        /// <summary>Insert completed successfully.</summary>
-        Completed,
-        /// <summary>OLC validation failed — caller should retry or fall back.</summary>
-        Restart,
-        /// <summary>Target leaf is full — needs pessimistic path for split/spill.</summary>
-        LeafFull,
-    }
-
-    /// <summary>Result of an OLC remove attempt.</summary>
-    private enum OlcRemoveResult
-    {
-        /// <summary>Remove completed successfully.</summary>
-        Completed,
-        /// <summary>OLC validation failed — caller should retry.</summary>
-        Restart,
-        /// <summary>Key not found in the tree (confirmed by OLC validation).</summary>
-        NotFound,
-        /// <summary>Remove requires merge/borrow or structural change — needs pessimistic path.</summary>
-        NeedsPessimistic,
-    }
-
     /// <summary>
     /// Spin-waits until the write lock is acquired. Counts contention spins for diagnostics.
     /// Returns true if lock was acquired immediately (no contention), false if spinning was needed.
     /// </summary>
+    /// <remarks>
+    /// Two-phase spin policy tuned for OLC latch hold times (~100-500 ns):
+    /// Phase 1: Tight PAUSE loop (64 iterations, ~100 ns on Zen / ~2 μs on Skylake+) — covers
+    ///          the common case of a leaf insert/remove completing on another core.
+    /// Phase 2: SpinWait with Sleep(1) disabled — escalates to Yield/Sleep(0) for rare splits/merges
+    ///          or SMT core-sharing, but never pays the 15 ms Windows timer-tick penalty.
+    /// </remarks>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     private bool SpinWriteLock(OlcLatch latch)
     {
@@ -1342,11 +1324,24 @@ public abstract partial class BTree<TKey> : IBTree where TKey : unmanaged
             return true; // no contention
         }
 
+        // Phase 1: tight PAUSE spin — stays on-core, covers typical latch hold time + cross-core coherence
+        for (int i = 0; i < 64; i++)
+        {
+            Interlocked.Increment(ref _writeLockFailures);
+            Thread.SpinWait(1);
+            if (latch.TryWriteLock())
+            {
+                return false; // contention detected
+            }
+        }
+
+        // Phase 2: yield-capped SpinWait — holder is likely doing a split/merge or sharing our core.
+        // sleep1Threshold: -1 disables Sleep(1) which would stall for ~15 ms (Windows timer tick).
         SpinWait spin = default;
         do
         {
             Interlocked.Increment(ref _writeLockFailures);
-            spin.SpinOnce();
+            spin.SpinOnce(sleep1Threshold: -1);
         }
         while (!latch.TryWriteLock());
         return false; // contention detected
@@ -1382,1365 +1377,6 @@ public abstract partial class BTree<TKey> : IBTree where TKey : unmanaged
         {
             _deferredLock.Exit(useMemoryBarrier: false);
         }
-    }
-
-    /// <summary>Creates the insert value, handling AllowMultiple buffer creation.</summary>
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private int CreateInsertValue(ref InsertArguments args, ref ChunkAccessor accessor)
-    {
-        if (AllowMultiple)
-        {
-            var bufferId = _storage.CreateBuffer(ref accessor);
-            args.ElementId = _storage.Append(bufferId, args.GetValue(), ref accessor);
-            args.BufferRootId = bufferId;
-            return bufferId;
-        }
-        return args.GetValue();
-    }
-
-    /// <summary>
-    /// OLC-dispatching remove: tries optimistic fast paths first, then falls back to pessimistic.
-    /// Begin/end remove fast paths and general mid-leaf remove operate without exclusive lock when the leaf has enough items (no merge/borrow needed).
-    /// All other cases use the pessimistic path.
-    /// </summary>
-    private void RemoveCore(ref RemoveArguments args)
-    {
-        if (IsEmpty())
-        {
-            return;
-        }
-
-        // OLC retry loop — handles begin/end fast paths + general non-merge removes
-        for (int attempt = 0; attempt < MaxOptimisticRestarts; attempt++)
-        {
-            var result = TryRemoveOlc(ref args);
-            if (result == OlcRemoveResult.Completed || result == OlcRemoveResult.NotFound)
-            {
-                return;
-            }
-            if (result == OlcRemoveResult.NeedsPessimistic)
-            {
-                break;
-            }
-            // Restart: continue loop
-            Interlocked.Increment(ref _optimisticRestarts);
-        }
-
-        // Pessimistic fallback
-        Interlocked.Increment(ref _pessimisticFallbacks);
-        RemoveCorePessimistic(ref args);
-    }
-
-    /// <summary>
-    /// OLC remove attempt: tries begin/end fast paths (first/last key of first/last leaf) and general mid-leaf remove via optimistic descent.
-    /// Only modifies a single leaf node (WriteLocked). Returns NeedsPessimistic when the leaf is too small (merge/borrow would be needed).
-    /// </summary>
-    private OlcRemoveResult TryRemoveOlc(ref RemoveArguments args)
-    {
-        ref var accessor = ref args.Accessor;
-
-        // --- Begin-remove fast path: remove first key of leftmost leaf ---
-        {
-            var ll = LinkList;
-            if (!ll.IsValid)
-            {
-                return OlcRemoveResult.Restart;
-            }
-            var llLatch = ll.GetLatch(ref accessor);
-            int llVersion = llLatch.ReadVersion();
-            if (llVersion == 0)
-            {
-                return OlcRemoveResult.Restart;
-            }
-
-            var firstKey = ll.GetFirst(ref accessor).Key;
-            if (!llLatch.ValidateVersion(llVersion))
-            {
-                return OlcRemoveResult.Restart;
-            }
-
-            int order = args.Compare(args.Key, firstKey);
-            if (order < 0)
-            {
-                return OlcRemoveResult.NotFound; // key < first key → definitely not in tree
-            }
-
-            if (order == 0)
-            {
-                int count = ll.GetCount(ref accessor);
-                bool isRoot = _rootChunkId == ll.ChunkId;
-                int capacity = ll.GetCapacity();
-                if (!llLatch.ValidateVersion(llVersion))
-                {
-                    return OlcRemoveResult.Restart;
-                }
-
-                // Safe if: root leaf with count > 1, or non-root leaf above half-full
-                if ((isRoot && count > 1) || (!isRoot && count > capacity / 2))
-                {
-                    if (!llLatch.TryWriteLock())
-                    {
-                        return OlcRemoveResult.Restart;
-                    }
-                    if (!llLatch.ValidateVersionLocked(llVersion))
-                    {
-                        llLatch.AbortWriteLock();
-                        return OlcRemoveResult.Restart;
-                    }
-                    // Re-verify first key under lock (concurrent OLC writer might have removed it)
-                    if (args.Compare(args.Key, ll.GetFirst(ref accessor).Key) != 0)
-                    {
-                        llLatch.WriteUnlock();
-                        return OlcRemoveResult.Restart;
-                    }
-
-                    args.SetRemovedValue(ll.PopFirstInternal(ref accessor).Value);
-                    llLatch.WriteUnlock();
-                    _hasCachedLastKey = false;
-                    DecCount();
-                    return OlcRemoveResult.Completed;
-                }
-
-                return OlcRemoveResult.NeedsPessimistic; // merge/borrow possible or tree might become empty
-            }
-        }
-
-        // --- End-remove fast path: remove last key of rightmost leaf ---
-        {
-            var rll = ReverseLinkList;
-            if (!rll.IsValid)
-            {
-                return OlcRemoveResult.Restart;
-            }
-            var rllLatch = rll.GetLatch(ref accessor);
-            int rllVersion = rllLatch.ReadVersion();
-            if (rllVersion == 0)
-            {
-                return OlcRemoveResult.Restart;
-            }
-
-            int rllCount = rll.GetCount(ref accessor);
-            if (rllCount == 0)
-            {
-                return OlcRemoveResult.Restart; // transient empty state during concurrent tree emptying
-            }
-            var lastKey = rll.GetItem(rllCount - 1, ref accessor).Key;
-            if (!rllLatch.ValidateVersion(rllVersion))
-            {
-                return OlcRemoveResult.Restart;
-            }
-
-            int order = args.Compare(args.Key, lastKey);
-            if (order > 0)
-            {
-                return OlcRemoveResult.NotFound; // key > last key → definitely not in tree
-            }
-
-            if (order == 0)
-            {
-                bool isRoot = _rootChunkId == rll.ChunkId;
-                int capacity = rll.GetCapacity();
-                if (!rllLatch.ValidateVersion(rllVersion))
-                {
-                    return OlcRemoveResult.Restart;
-                }
-
-                if ((isRoot && rllCount > 1) || (!isRoot && rllCount > capacity / 2))
-                {
-                    if (!rllLatch.TryWriteLock())
-                    {
-                        return OlcRemoveResult.Restart;
-                    }
-                    if (!rllLatch.ValidateVersionLocked(rllVersion))
-                    {
-                        rllLatch.AbortWriteLock();
-                        return OlcRemoveResult.Restart;
-                    }
-                    // Re-verify last key under lock
-                    if (args.Compare(args.Key, rll.GetLast(ref accessor).Key) != 0)
-                    {
-                        rllLatch.WriteUnlock();
-                        return OlcRemoveResult.Restart;
-                    }
-
-                    args.SetRemovedValue(rll.PopLastInternal(ref accessor).Value);
-                    rllLatch.WriteUnlock();
-                    _hasCachedLastKey = false;
-                    DecCount();
-                    return OlcRemoveResult.Completed;
-                }
-
-                return OlcRemoveResult.NeedsPessimistic;
-            }
-        }
-
-        // --- General path: optimistic descent to leaf, remove if safe (no merge/borrow) ---
-        var (leafChunkId, leafVersion, keyIndex) = OptimisticDescendToLeaf(args.Key, ref accessor);
-        if (leafChunkId == 0)
-        {
-            return OlcRemoveResult.Restart;
-        }
-
-        if (keyIndex < 0)
-        {
-            // Key not found — validate version to confirm
-            var nfLeaf = new NodeWrapper(_storage, leafChunkId);
-            if (!nfLeaf.GetLatch(ref accessor).ValidateVersion(leafVersion))
-            {
-                return OlcRemoveResult.Restart;
-            }
-            return OlcRemoveResult.NotFound;
-        }
-
-        // Key found — check if safe to remove under OLC (no merge/borrow needed)
-        {
-            var leaf = new NodeWrapper(_storage, leafChunkId);
-            var leafLatch = leaf.GetLatch(ref accessor);
-            int count = leaf.GetCount(ref accessor);
-            bool isRoot = _rootChunkId == leafChunkId;
-            int capacity = leaf.GetCapacity();
-            if (!leafLatch.ValidateVersion(leafVersion))
-            {
-                return OlcRemoveResult.Restart;
-            }
-
-            if ((isRoot && count > 1) || (!isRoot && count > capacity / 2))
-            {
-                if (!leafLatch.TryWriteLock())
-                {
-                    return OlcRemoveResult.Restart;
-                }
-                if (!leafLatch.ValidateVersionLocked(leafVersion))
-                {
-                    leafLatch.AbortWriteLock();
-                    return OlcRemoveResult.Restart;
-                }
-
-                // Re-find key under lock (index might have shifted due to concurrent modification)
-                var reIndex = leaf.Find(args.Key, args.Comparer, ref accessor);
-                if (reIndex < 0)
-                {
-                    leafLatch.WriteUnlock();
-                    return OlcRemoveResult.NotFound; // concurrent writer already removed it
-                }
-
-                args.SetRemovedValue(leaf.RemoveAtInternal(reIndex, ref accessor).Value);
-                leafLatch.WriteUnlock();
-                _hasCachedLastKey = false;
-                DecCount();
-                return OlcRemoveResult.Completed;
-            }
-
-            return OlcRemoveResult.NeedsPessimistic;
-        }
-    }
-
-    /// <summary>
-    /// Pessimistic remove fallback: uses WriteLock/WriteUnlock on individual nodes so concurrent OLC readers detect changes. No global lock — concurrency is
-    /// handled by per-node OLC latches and latch-coupled SMO in RemoveIterative.
-    /// </summary>
-    private void RemoveCorePessimistic(ref RemoveArguments args)
-    {
-        ref var accessor = ref args.Accessor;
-        try
-        {
-            if (IsEmpty())
-            {
-                return;
-            }
-
-            // Begin-remove fast path (WriteLock protects against concurrent OLC writers)
-            {
-                var ll = LinkList;
-                SpinWriteLock(ll.GetLatch(ref accessor));
-                int order = args.Compare(args.Key, ll.GetFirst(ref accessor).Key);
-                if (order < 0)
-                {
-                    ll.GetLatch(ref accessor).AbortWriteLock(); // key not in tree — didn't modify node
-                    return;
-                }
-
-                if (order == 0 && (Root == ll || ll.GetCount(ref accessor) > ll.GetCapacity() / 2))
-                {
-                    args.SetRemovedValue(ll.PopFirstInternal(ref accessor).Value);
-                    ll.GetLatch(ref accessor).WriteUnlock();
-                    _hasCachedLastKey = false;
-                    DecCount();
-                    if (IsEmpty())
-                    {
-                        Root = LinkList = ReverseLinkList = default;
-                        Height--;
-                    }
-                    return;
-                }
-                ll.GetLatch(ref accessor).AbortWriteLock(); // condition failed — didn't modify node
-            }
-
-            // End-remove fast path
-            {
-                var rll = ReverseLinkList;
-                SpinWriteLock(rll.GetLatch(ref accessor));
-
-                // Safety: if rll was split concurrently, it's no longer the rightmost leaf.
-                // Fall through to general path which handles stale pointers correctly.
-                if (rll.GetNext(ref accessor).IsValid)
-                {
-                    rll.GetLatch(ref accessor).AbortWriteLock();
-                }
-                else
-                {
-                    int order = args.Compare(args.Key, rll.GetLast(ref accessor).Key);
-                    if (order > 0)
-                    {
-                        rll.GetLatch(ref accessor).AbortWriteLock(); // key not in tree — didn't modify node
-                        return;
-                    }
-
-                    if (order == 0 && (Root == rll || rll.GetCount(ref accessor) > rll.GetCapacity() / 2))
-                    {
-                        args.SetRemovedValue(rll.PopLastInternal(ref accessor).Value);
-                        rll.GetLatch(ref accessor).WriteUnlock();
-                        _hasCachedLastKey = false;
-                        DecCount();
-                        return;
-                    }
-                    rll.GetLatch(ref accessor).AbortWriteLock(); // condition failed — didn't modify node
-                }
-            }
-
-            // General remove path with latch-coupled SMO — retry on lock contention
-            _hasCachedLastKey = false;
-            bool merge;
-            SpinWait spin = default;
-            while (true)
-            {
-                merge = RemoveIterative(ref args, ref accessor, out bool removeCompleted);
-                if (removeCompleted)
-                {
-                    break;
-                }
-                Interlocked.Increment(ref _optimisticRestarts);
-                spin.SpinOnce();
-            }
-
-            if (args.Removed)
-            {
-                DecCount();
-            }
-
-            if (merge && Root.GetLength(ref accessor) == 0)
-            {
-                Root = Root.GetChild(-1, ref accessor); // left most child becomes root. (returns null for leafs)
-                if (Root.IsValid == false)
-                {
-                    LinkList = default;
-                    ReverseLinkList = default;
-                }
-                Height--;
-            }
-
-            if (ReverseLinkList.IsValid && ReverseLinkList.GetPrevious(ref accessor).IsValid && ReverseLinkList.GetPrevious(ref accessor).GetNext(ref accessor).IsValid == false)
-            {
-                ReverseLinkList = ReverseLinkList.GetPrevious(ref accessor);
-            }
-        }
-        finally
-        {
-            // Reclaim deferred nodes whose epoch is safe (all readers have exited).
-            DeferredReclaim();
-        }
-    }
-
-    private void AddOrUpdateCore(ref InsertArguments args)
-    {
-        ref var accessor = ref args.Accessor;
-
-        // 1. Empty tree initialization.
-        //    An empty tree has no root node, so OLC readers/writers have nothing to latch on.
-        //    CAS on _rootChunkId atomically races to claim the init slot; loser sees non-zero root and proceeds.
-        if (IsEmpty())
-        {
-            var newRoot = AllocNode(NodeStates.IsLeaf, ref accessor);
-            if (Interlocked.CompareExchange(ref _rootChunkId, newRoot.ChunkId, 0) == 0)
-            {
-                // We won the race — initialize root, LinkList, ReverseLinkList
-                LinkList = newRoot;
-                ReverseLinkList = newRoot;
-                Height++;
-                int value = CreateInsertValue(ref args, ref accessor);
-                newRoot.PushLast(new KeyValueItem(args.Key, value), ref accessor);
-                IncCount();
-                _cachedLastKey = args.Key;
-                _hasCachedLastKey = true;
-                return;
-            }
-            // Another thread initialized the root — free our unused node and fall through
-            _segment.FreeChunk(newRoot.ChunkId);
-        }
-
-        // 2. OLC retry loop — handles append/prepend fast paths + non-full leaf inserts.
-        //    Zero writes to shared state except the single leaf being modified (WriteLocked).
-        for (int attempt = 0; attempt < MaxOptimisticRestarts; attempt++)
-        {
-            var result = TryInsertOlc(ref args);
-            if (result == OlcInsertResult.Completed)
-            {
-                return;
-            }
-            if (result == OlcInsertResult.LeafFull)
-            {
-                break; // Need pessimistic path for split/spill
-            }
-            // Restart: version validation failed
-            Interlocked.Increment(ref _optimisticRestarts);
-        }
-
-        // 3. Pessimistic fallback — exclusive lock + WriteLock all modified nodes for OLC readers
-        Interlocked.Increment(ref _pessimisticFallbacks);
-        AddOrUpdateCorePessimistic(ref args);
-    }
-
-    /// <summary>
-    /// OLC insert attempt: tries append/prepend fast paths and non-full leaf insert.
-    /// Only modifies a single leaf node (WriteLocked). Returns LeafFull when the target leaf is full and needs split/spill (which requires the pessimistic path).
-    /// </summary>
-    private OlcInsertResult TryInsertOlc(ref InsertArguments args)
-    {
-        ref var accessor = ref args.Accessor;
-
-        // --- Append fast path: insert at end of rightmost leaf ---
-        var rl = ReverseLinkList;
-        if (rl.IsValid)
-        {
-            var latch = rl.GetLatch(ref accessor);
-            var version = latch.ReadVersion();
-            if (version != 0)
-            {
-                int rlCount = rl.GetCount(ref accessor);
-                if (rlCount > 0)
-                {
-                    var lastKey = rl.GetLast(ref accessor).Key;
-                    if (!latch.ValidateVersion(version))
-                    {
-                        return OlcInsertResult.Restart;
-                    }
-
-                    int order = args.Compare(args.Key, lastKey);
-                    if (order > 0)
-                    {
-                        if (!latch.TryWriteLock())
-                        {
-                            return OlcInsertResult.Restart;
-                        }
-                        if (!latch.ValidateVersionLocked(version))
-                        {
-                            latch.AbortWriteLock();
-                            return OlcInsertResult.Restart;
-                        }
-                        // Safety: if rl was split concurrently, it's no longer the rightmost leaf.
-                        // GetNext().IsValid means a new right sibling exists — abort and fall through.
-                        if (rl.GetIsFull(ref accessor) || rl.GetNext(ref accessor).IsValid)
-                        {
-                            latch.AbortWriteLock();
-                            return rl.GetIsFull(ref accessor) ? OlcInsertResult.LeafFull : OlcInsertResult.Restart;
-                        }
-                        int value = CreateInsertValue(ref args, ref accessor);
-                        rl.PushLast(new KeyValueItem(args.Key, value), ref accessor);
-                        _cachedLastKey = args.Key;
-                        _hasCachedLastKey = true;
-                        latch.WriteUnlock();
-                        IncCount();
-                        return OlcInsertResult.Completed;
-                    }
-
-                    if (order == 0)
-                    {
-                        if (AllowMultiple)
-                        {
-                            if (!latch.TryWriteLock())
-                            {
-                                return OlcInsertResult.Restart;
-                            }
-                            if (!latch.ValidateVersionLocked(version))
-                            {
-                                latch.AbortWriteLock();
-                                return OlcInsertResult.Restart;
-                            }
-                            var bufferRootId = rl.GetLast(ref accessor).Value;
-                            args.ElementId = _storage.Append(bufferRootId, args.GetValue(), ref accessor);
-                            args.BufferRootId = bufferRootId;
-                            latch.WriteUnlock();
-                            return OlcInsertResult.Completed;
-                        }
-                        ThrowHelper.ThrowUniqueConstraintViolation();
-                    }
-                }
-            }
-        }
-
-        // --- Prepend fast path: insert at beginning of leftmost leaf ---
-        var ll = LinkList;
-        if (ll.IsValid)
-        {
-            var llLatch = ll.GetLatch(ref accessor);
-            var llVersion = llLatch.ReadVersion();
-            if (llVersion != 0)
-            {
-                int llCount = ll.GetCount(ref accessor);
-                if (llCount > 0)
-                {
-                    var firstKey = ll.GetFirst(ref accessor).Key;
-                    if (!llLatch.ValidateVersion(llVersion))
-                    {
-                        return OlcInsertResult.Restart;
-                    }
-
-                    int order = args.Compare(args.Key, firstKey);
-                    if (order < 0)
-                    {
-                        if (!llLatch.TryWriteLock())
-                        {
-                            return OlcInsertResult.Restart;
-                        }
-                        if (!llLatch.ValidateVersionLocked(llVersion))
-                        {
-                            llLatch.AbortWriteLock();
-                            return OlcInsertResult.Restart;
-                        }
-                        if (ll.GetIsFull(ref accessor))
-                        {
-                            llLatch.WriteUnlock();
-                            return OlcInsertResult.LeafFull;
-                        }
-                        int value = CreateInsertValue(ref args, ref accessor);
-                        ll.PushFirst(new KeyValueItem(args.Key, value), ref accessor);
-                        llLatch.WriteUnlock();
-                        IncCount();
-                        return OlcInsertResult.Completed;
-                    }
-
-                    if (order == 0)
-                    {
-                        if (AllowMultiple)
-                        {
-                            if (!llLatch.TryWriteLock())
-                            {
-                                return OlcInsertResult.Restart;
-                            }
-                            if (!llLatch.ValidateVersionLocked(llVersion))
-                            {
-                                llLatch.AbortWriteLock();
-                                return OlcInsertResult.Restart;
-                            }
-                            var bufferRootId = ll.GetFirst(ref accessor).Value;
-                            args.ElementId = _storage.Append(bufferRootId, args.GetValue(), ref accessor);
-                            args.BufferRootId = bufferRootId;
-                            llLatch.WriteUnlock();
-                            return OlcInsertResult.Completed;
-                        }
-                        ThrowHelper.ThrowUniqueConstraintViolation();
-                    }
-                }
-            }
-        }
-
-        // --- General path: optimistic descent to leaf, non-full insert ---
-        // followRightLink: false — inserts must not follow B-link because inserting into the right sibling bypasses the spill/split path that updates parent
-        // separators. If the leaf was split concurrently, version validation will trigger a restart.
-        var (leafChunkId, leafVersion, keyIndex) = OptimisticDescendToLeaf(args.Key, ref accessor, followRightLink: false);
-        if (leafChunkId == 0)
-        {
-            return OlcInsertResult.Restart;
-        }
-
-        var leaf = new NodeWrapper(_storage, leafChunkId);
-        var leafLatch = leaf.GetLatch(ref accessor);
-        if (!leafLatch.TryWriteLock())
-        {
-            return OlcInsertResult.Restart;
-        }
-        if (!leafLatch.ValidateVersionLocked(leafVersion))
-        {
-            leafLatch.AbortWriteLock();
-            return OlcInsertResult.Restart;
-        }
-        // Range check: stale separator may route to wrong leaf after a concurrent split.
-        // High key is an exclusive upper bound, so key >= highKey means we're out of range.
-        if (leaf.GetCount(ref accessor) > 0 && leaf.GetNext(ref accessor).IsValid && args.Compare(args.Key, leaf.GetHighKey(ref accessor)) >= 0)
-        {
-            leafLatch.WriteUnlock();
-            return OlcInsertResult.Restart;
-        }
-        if (leaf.GetIsFull(ref accessor))
-        {
-            leafLatch.WriteUnlock();
-            return OlcInsertResult.LeafFull;
-        }
-
-        // Re-search under lock (key positions may have shifted since optimistic read)
-        keyIndex = leaf.Find(args.Key, args.KeyComparer, ref accessor);
-        if (keyIndex < 0)
-        {
-            keyIndex = ~keyIndex;
-            int value = CreateInsertValue(ref args, ref accessor);
-            leaf.Insert(keyIndex, new KeyValueItem(args.Key, value), ref accessor);
-            leafLatch.WriteUnlock();
-            IncCount();
-            return OlcInsertResult.Completed;
-        }
-
-        // Key already exists
-        if (AllowMultiple)
-        {
-            var curItem = leaf.GetItem(keyIndex, ref accessor);
-            args.ElementId = _storage.Append(curItem.Value, args.GetValue(), ref accessor);
-            args.BufferRootId = curItem.Value;
-            leafLatch.WriteUnlock();
-            return OlcInsertResult.Completed;
-        }
-        leafLatch.WriteUnlock();
-        ThrowHelper.ThrowUniqueConstraintViolation();
-        return OlcInsertResult.Restart; // unreachable — ThrowHelper always throws
-    }
-
-    /// <summary>
-    /// Pessimistic insert fallback: uses InsertIterative with latch-coupled SMO.
-    /// No global lock — concurrency is handled by per-node OLC latches.
-    /// </summary>
-    private void AddOrUpdateCorePessimistic(ref InsertArguments args)
-    {
-        try
-        {
-            ref var accessor = ref args.Accessor;
-
-            if (IsEmpty())
-            {
-                Root = AllocNode(NodeStates.IsLeaf, ref accessor);
-                LinkList = Root;
-                ReverseLinkList = LinkList;
-                Height++;
-            }
-
-            // Append fast path: lock the last leaf and insert if key > lastKey and leaf not full.
-            // Bypass when leaf is contended and sufficiently populated — fall through to InsertIterative
-            // which has the path recording needed for contention split propagation.
-            // Capture local refs to avoid races from concurrent ReverseLinkList/LinkList updates.
-            {
-                var rl = ReverseLinkList;
-                var bypassAppendFastPath = rl.IsValid && rl.GetContentionHint(ref accessor) >= ContentionSplitThreshold && rl.GetCount(ref accessor) > rl.GetCapacity() / 2;
-                var order = IsEmpty() ? 1 : args.Compare(args.Key, _hasCachedLastKey ? _cachedLastKey : rl.GetLast(ref accessor).Key);
-                if (!bypassAppendFastPath && order > 0 && !rl.GetIsFull(ref accessor))
-                {
-                    var rlLatch = rl.GetLatch(ref accessor);
-                    SpinWriteLock(rlLatch);
-                    // Re-validate under lock: leaf may now be full, another writer inserted a larger key,
-                    // or a concurrent split made this leaf no longer the rightmost (GetNext becomes valid).
-                    if (!rl.GetIsFull(ref accessor) && !rl.GetNext(ref accessor).IsValid && args.Compare(args.Key, rl.GetLast(ref accessor).Key) > 0)
-                    {
-                        int value = CreateInsertValue(ref args, ref accessor);
-                        rl.PushLast(new KeyValueItem(args.Key, value), ref accessor);
-                        _cachedLastKey = args.Key;
-                        _hasCachedLastKey = true;
-                        rlLatch.WriteUnlock();
-                        IncCount();
-                        return;
-                    }
-                    rlLatch.AbortWriteLock();
-                    // Fall through to general path
-                }
-                else if (order == 0 && AllowMultiple)
-                {
-                    var rlLatch = rl.GetLatch(ref accessor);
-                    SpinWriteLock(rlLatch);
-                    var lastEntry = rl.GetLast(ref accessor);
-                    if (args.Compare(args.Key, lastEntry.Key) == 0)
-                    {
-                        args.ElementId = _storage.Append(lastEntry.Value, args.GetValue(), ref accessor);
-                        args.BufferRootId = lastEntry.Value;
-                        rlLatch.WriteUnlock();
-                        return;
-                    }
-                    rlLatch.AbortWriteLock();
-                    // Fall through
-                }
-                else if (order == 0)
-                {
-                    ThrowHelper.ThrowUniqueConstraintViolation();
-                }
-            }
-
-            // Prepend fast path: lock the first leaf and insert if key < firstKey and leaf not full.
-            if (!IsEmpty())
-            {
-                var ll = LinkList;
-                int order = args.Compare(args.Key, ll.GetFirst(ref accessor).Key);
-                if (order < 0 && !ll.GetIsFull(ref accessor))
-                {
-                    var llLatch = ll.GetLatch(ref accessor);
-                    SpinWriteLock(llLatch);
-                    if (!ll.GetIsFull(ref accessor) && args.Compare(args.Key, ll.GetFirst(ref accessor).Key) < 0)
-                    {
-                        int value = CreateInsertValue(ref args, ref accessor);
-                        ll.PushFirst(new KeyValueItem(args.Key, value), ref accessor);
-                        llLatch.WriteUnlock();
-                        IncCount();
-                        return;
-                    }
-                    llLatch.AbortWriteLock();
-                    // Fall through
-                }
-                else if (order == 0 && AllowMultiple)
-                {
-                    var llLatch = ll.GetLatch(ref accessor);
-                    SpinWriteLock(llLatch);
-                    var firstEntry = ll.GetFirst(ref accessor);
-                    if (args.Compare(args.Key, firstEntry.Key) == 0)
-                    {
-                        args.ElementId = _storage.Append(firstEntry.Value, args.GetValue(), ref accessor);
-                        args.BufferRootId = firstEntry.Value;
-                        llLatch.WriteUnlock();
-                        return;
-                    }
-                    llLatch.AbortWriteLock();
-                    // Fall through
-                }
-                else if (order == 0)
-                {
-                    ThrowHelper.ThrowUniqueConstraintViolation();
-                }
-            }
-
-            // General path with latch-coupled SMO — retry on lock contention
-            // InsertIterative handles root splits internally under the root's write lock.
-            SpinWait spin = default;
-            while (true)
-            {
-                InsertIterative(ref args, ref accessor, out bool insertCompleted);
-                if (insertCompleted)
-                {
-                    break;
-                }
-                Interlocked.Increment(ref _optimisticRestarts);
-                spin.SpinOnce();
-            }
-
-            if (args.Added)
-            {
-                IncCount();
-            }
-            else if (!AllowMultiple)
-            {
-                ThrowHelper.ThrowUniqueConstraintViolation();
-            }
-
-            var next = ReverseLinkList.GetNext(ref accessor);
-            if (next.IsValid)
-            {
-                ReverseLinkList = next;
-            }
-            _cachedLastKey = GetLast(ref accessor).Key;
-            _hasCachedLastKey = true;
-        }
-        finally
-        {
-            // Reclaim deferred nodes whose epoch is safe (all readers have exited).
-            DeferredReclaim();
-        }
-    }
-
-    /// <summary>
-    /// Iterative insert with latch-coupled SMO: descends optimistically recording PathVersions, then locks bottom-up only as needed for structural modifications.
-    /// Fast path (leaf not full): locks only the leaf node.
-    /// Slow path (leaf full, new key): locks leaf + neighbors + path nodes with version validation.
-    /// Returns null if no root split, non-null promoted key if root split needed.
-    /// Sets <paramref name="completed"/> to false when lock acquisition fails and caller must retry.
-    /// </summary>
-    private KeyValueItem? InsertIterative(ref InsertArguments args, ref ChunkAccessor accessor, out bool completed)
-    {
-        completed = false;
-        MutationContext ctx = default;
-        var node = Root;
-        var relatives = new NodeRelatives();
-
-        // Phase 1: Descend from root to leaf, recording path + PathVersions for validation.
-        // OLC protocol: read version BEFORE data, validate AFTER — ensures (index, version) are consistent.
-        while (!node.GetIsLeaf(ref accessor))
-        {
-            var latch = node.GetLatch(ref accessor);
-            int version = latch.ReadVersion();
-            if (version == 0)
-            {
-                return null; // node locked or obsolete — restart
-            }
-
-            var index = node.Find(args.Key, args.KeyComparer, ref accessor);
-            if (index < 0)
-            {
-                index = ~index - 1;
-            }
-
-            var child = node.GetChild(index, ref accessor);
-            int parentCount = node.GetCount(ref accessor);
-
-            // Validate: node wasn't modified during our unlocked read
-            if (!latch.ValidateVersion(version))
-            {
-                return null; // node modified between version read and data read — restart
-            }
-
-            NodeRelatives.Create(child, index, node, parentCount, ref relatives, out var childRelatives, ref accessor);
-
-            ctx.PathNodes[ctx.Depth] = node;
-            ctx.PathChildIndices[ctx.Depth] = index;
-            ctx.PathVersions[ctx.Depth] = version;
-
-            // Store after Create so lazy-resolved siblings are cached in the stored copy
-            ctx.PathRelatives[ctx.Depth] = relatives;
-
-            node = child;
-            relatives = childRelatives;
-            ctx.Depth++;
-        }
-
-        // Phase 1.5A: Lock leaf with version validation.
-        // Between Phase 1 descent and lock acquisition, a concurrent writer may have split/modified this leaf. Snapshot the version before locking,
-        // then validate after.
-        var leafLatch = node.GetLatch(ref accessor);
-        int leafVersion = leafLatch.ReadVersion();
-        if (leafVersion == 0)
-        {
-            // Leaf is locked or obsolete. SpinWriteLock to wait for the current holder to release, then restart — we can't validate without a baseline version.
-            SpinWriteLock(leafLatch);
-            leafLatch.AbortWriteLock(); // release without version bump (we didn't modify anything)
-            return null;
-        }
-        bool leafAcquiredClean = SpinWriteLock(leafLatch);
-        if (!leafLatch.ValidateVersionLocked(leafVersion))
-        {
-            leafLatch.AbortWriteLock(); // release without version bump — leaf was modified, not by us
-            return null; // restart — leaf was modified between descent and lock
-        }
-
-        // Update contention hint: saturating counter for detecting hot leaves
-        {
-            int hint = node.GetContentionHint(ref accessor);
-            if (!leafAcquiredClean)
-            {
-                node.SetContentionHint(Math.Min(hint + 1, 255), ref accessor);
-            }
-            else if (hint > 0)
-            {
-                node.SetContentionHint(hint - 1, ref accessor);
-            }
-        }
-
-        // B-link move_right (Lehman & Yao): if the key is beyond this leaf's range, a concurrent split moved some keys to a right sibling. Chain right using
-        // lock coupling (lock next before releasing current) until we find the correct leaf. Forward progress is guaranteed:
-        // all movement is strictly rightward with no cycle, and SpinWriteLock waits for busy siblings.
-        bool movedRight = false;
-        // High key is an exclusive upper bound, so key >= highKey means we're out of range.
-        while (node.GetCount(ref accessor) > 0 && node.GetNext(ref accessor).IsValid && args.Compare(args.Key, node.GetHighKey(ref accessor)) >= 0)
-        {
-            Interlocked.Increment(ref _moveRightCount);
-            var nextNode = node.GetNext(ref accessor);
-            SpinWriteLock(nextNode.GetLatch(ref accessor));
-
-            // Gap check: after locking next leaf, verify key belongs there.
-            // Without this, move_right chains across subtree boundaries when the key space has gaps (e.g., leaves [14-26] → [201-213] with no intermediate leaves).
-            // Key 27 would land in [201-213] where it doesn't belong → BST violation.
-            if (nextNode.GetCount(ref accessor) > 0 &&
-                args.Compare(args.Key, nextNode.GetFirst(ref accessor).Key) < 0)
-            {
-                nextNode.GetLatch(ref accessor).AbortWriteLock();
-                break; // Key falls in a gap — belongs in current leaf (will split via slow path)
-            }
-
-            node.GetLatch(ref accessor).AbortWriteLock();    // release current
-            node = nextNode;
-            movedRight = true;
-        }
-
-        // Fast path: leaf not full → InsertLeaf only modifies this leaf (insert or duplicate append)
-        // If contention is high and leaf is sufficiently populated, fall through to contention split.
-        bool itemAlreadyInserted = false;
-        if (!node.GetIsFull(ref accessor))
-        {
-            node.InsertLeaf(ref args, ref relatives, ref accessor);
-            itemAlreadyInserted = true;
-
-            bool shouldContentionSplit = !movedRight && node.GetContentionHint(ref accessor) >= ContentionSplitThreshold
-                                                     && node.GetCount(ref accessor) > node.GetCapacity() / 2;
-
-            if (!shouldContentionSplit)
-            {
-                node.GetLatch(ref accessor).WriteUnlock();
-                completed = true;
-                return null; // no split needed
-            }
-
-            // Contention split: reset hint and fall through to split path
-            node.SetContentionHint(0, ref accessor);
-            Interlocked.Increment(ref _contentionSplitCount);
-            // Fall through to split path below
-        }
-
-        // Check if key already exists in full leaf (buffer append, no structural change)
-        if (!itemAlreadyInserted)
-        {
-            var idx = node.Find(args.Key, args.KeyComparer, ref accessor);
-            if (idx >= 0)
-            {
-                node.InsertLeaf(ref args, ref relatives, ref accessor);
-                node.GetLatch(ref accessor).WriteUnlock();
-                completed = true;
-                return null;
-            }
-        }
-
-        // After move_right, PathVersions and relatives are stale (recorded for the original leaf's// path). The gap check above prevents wrong-subtree inserts.
-        // Force split (skip spill which uses stale relatives) and don't propagate — B-link right-link chain provides correct routing until a future insert
-        // naturally propagates the separator.
-        // This avoids the restart livelock that occurs when the slow path repeatedly fails.
-        if (movedRight)
-        {
-            // Lock the next neighbor for the split's linked list update (SetPrevious on next node)
-            var mrNext = node.GetNext(ref accessor);
-            if (mrNext.IsValid)
-            {
-                SpinWriteLock(mrNext.GetLatch(ref accessor));
-            }
-
-            node.InsertLeaf(ref args, ref relatives, ref accessor, forceSplit: true);
-
-            if (mrNext.IsValid)
-            {
-                mrNext.GetLatch(ref accessor).WriteUnlock();
-            }
-            node.GetLatch(ref accessor).WriteUnlock();
-            completed = true;
-            return null; // promoted key not propagated — B-link covers the gap
-        }
-
-        // Slow path: leaf full or contention split — structural modification needed.
-        // For contention split, skip leafPrev lock (no spill needed — item already in, only need right neighbor for linked list).
-        // On lock failure: contention split uses WriteUnlock + completed=true (item is in); regular uses AbortWriteLock + restart.
-        var leafPrev = itemAlreadyInserted ? default : node.GetPrevious(ref accessor);
-        var leafNext = node.GetNext(ref accessor);
-        if (leafPrev.IsValid && !leafPrev.GetLatch(ref accessor).TryWriteLock())
-        {
-            node.GetLatch(ref accessor).AbortWriteLock();
-            return null; // restart
-        }
-        if (leafNext.IsValid && !leafNext.GetLatch(ref accessor).TryWriteLock())
-        {
-            if (leafPrev.IsValid)
-            {
-                leafPrev.GetLatch(ref accessor).AbortWriteLock();
-            }
-            if (itemAlreadyInserted)
-            {
-                // Contention split abort: item is already inserted, just skip the proactive split
-                node.GetLatch(ref accessor).WriteUnlock();
-                completed = true;
-                return null;
-            }
-            node.GetLatch(ref accessor).AbortWriteLock();
-            return null; // restart
-        }
-
-        // Lock path nodes bottom-up with version validation.
-        // Required for ancestor key updates during spill and split propagation.
-        for (int i = ctx.Depth - 1; i >= 0; i--)
-        {
-            var pathLatch = ctx.PathNodes[i].GetLatch(ref accessor);
-            if (!pathLatch.TryWriteLock())
-            {
-                // Unlock path nodes already acquired above this level
-                for (int j = i + 1; j < ctx.Depth; j++)
-                {
-                    ctx.PathNodes[j].GetLatch(ref accessor).AbortWriteLock();
-                }
-                if (leafNext.IsValid)
-                {
-                    leafNext.GetLatch(ref accessor).AbortWriteLock();
-                }
-                if (leafPrev.IsValid)
-                {
-                    leafPrev.GetLatch(ref accessor).AbortWriteLock();
-                }
-                if (itemAlreadyInserted)
-                {
-                    node.GetLatch(ref accessor).WriteUnlock();
-                    completed = true;
-                    return null;
-                }
-                node.GetLatch(ref accessor).AbortWriteLock();
-                return null; // restart
-            }
-            if (!pathLatch.ValidateVersionLocked(ctx.PathVersions[i]))
-            {
-                pathLatch.AbortWriteLock();
-                for (int j = i + 1; j < ctx.Depth; j++)
-                {
-                    ctx.PathNodes[j].GetLatch(ref accessor).AbortWriteLock();
-                }
-                if (leafNext.IsValid)
-                {
-                    leafNext.GetLatch(ref accessor).AbortWriteLock();
-                }
-                if (leafPrev.IsValid)
-                {
-                    leafPrev.GetLatch(ref accessor).AbortWriteLock();
-                }
-                if (itemAlreadyInserted)
-                {
-                    node.GetLatch(ref accessor).WriteUnlock();
-                    completed = true;
-                    return null;
-                }
-                node.GetLatch(ref accessor).AbortWriteLock();
-                return null; // restart
-            }
-        }
-
-        // All needed nodes locked — Phase 2: Insert at leaf (may spill or split) or contention split
-        KeyValueItem? promoted;
-        if (itemAlreadyInserted)
-        {
-            // Contention split: item is already in the leaf, just redistribute via SplitLeafRight
-            var rightNode = node.SplitLeafRight(ref accessor);
-            promoted = new KeyValueItem(rightNode.GetFirst(ref accessor).Key, rightNode.ChunkId);
-        }
-        else
-        {
-            promoted = node.InsertLeaf(ref args, ref relatives, ref accessor);
-        }
-
-        // Phase 2.5: Unlock leaf neighbors (version bumped by WriteUnlock)
-        if (leafNext.IsValid)
-        {
-            leafNext.GetLatch(ref accessor).WriteUnlock();
-        }
-        if (leafPrev.IsValid)
-        {
-            leafPrev.GetLatch(ref accessor).WriteUnlock();
-        }
-        // Defer leaf unlock if this is a root-leaf that split (need to hold lock for atomic root creation)
-        if (!(ctx.Depth == 0 && promoted != null))
-        {
-            node.GetLatch(ref accessor).WriteUnlock();
-        }
-
-        // Phase 3: Propagate splits upward through internal nodes
-        while (ctx.Depth > 0 && promoted != null)
-        {
-            ctx.Depth--;
-            node = ctx.PathNodes[ctx.Depth];
-            relatives = ctx.PathRelatives[ctx.Depth];
-
-            // Lock siblings that HandlePromotedInsert might spill to (only when node is full)
-            NodeWrapper leftSib = default, rightSib = default;
-            if (node.GetIsFull(ref accessor))
-            {
-                leftSib = relatives.GetLeftSibling(ref accessor);
-                rightSib = relatives.GetRightSibling(ref accessor);
-                if (leftSib.IsValid)
-                {
-                    SpinWriteLock(leftSib.GetLatch(ref accessor));
-                }
-                if (rightSib.IsValid)
-                {
-                    SpinWriteLock(rightSib.GetLatch(ref accessor));
-                }
-            }
-
-            promoted = node.HandlePromotedInsert(ctx.PathChildIndices[ctx.Depth], promoted.Value, ref relatives, ref accessor);
-
-            // Unlock siblings
-            if (rightSib.IsValid)
-            {
-                rightSib.GetLatch(ref accessor).WriteUnlock();
-            }
-            if (leftSib.IsValid)
-            {
-                leftSib.GetLatch(ref accessor).WriteUnlock();
-            }
-            // Defer root unlock if root split (need to hold lock for atomic root creation)
-            if (!(ctx.Depth == 0 && promoted != null))
-            {
-                node.GetLatch(ref accessor).WriteUnlock();
-            }
-        }
-
-        // Phase 3.5: Unlock remaining path nodes above propagation level
-        while (ctx.Depth > 0)
-        {
-            ctx.Depth--;
-            ctx.PathNodes[ctx.Depth].GetLatch(ref accessor).WriteUnlock();
-        }
-
-        // Phase 4: Root split — create new root while holding old root's write lock.
-        // This prevents concurrent InsertIterative calls from racing to create multiple roots.
-        if (promoted != null)
-        {
-            var newRoot = AllocNode(NodeStates.None, ref accessor);
-            newRoot.SetLeft(Root, ref accessor);
-            newRoot.Insert(0, promoted.Value, ref accessor);
-            Root = newRoot;
-            Height++;
-            node.GetLatch(ref accessor).WriteUnlock(); // release old root after publishing new root
-        }
-
-        completed = true;
-        return null; // root splits handled internally — never return promoted to caller
-    }
-
-    /// <summary>
-    /// Iterative remove with latch-coupled SMO: descends optimistically recording PathVersions, then locks bottom-up only as needed for structural modifications.
-    /// Fast path (leaf stays half-full or root leaf): locks only the leaf node.
-    /// Slow path (leaf underflows): locks leaf + neighbors + path nodes with version validation.
-    /// Sets <paramref name="completed"/> to false when lock acquisition fails and caller must retry.
-    /// </summary>
-    private bool RemoveIterative(ref RemoveArguments args, ref ChunkAccessor accessor, out bool completed)
-    {
-        completed = false;
-        MutationContext ctx = default;
-        var node = Root;
-        var relatives = new NodeRelatives();
-
-        // Phase 1: Descend from root to leaf, recording path + PathVersions for validation.
-        // OLC protocol: read version BEFORE data, validate AFTER — ensures (index, version) are consistent.
-        while (!node.GetIsLeaf(ref accessor))
-        {
-            var latch = node.GetLatch(ref accessor);
-            int version = latch.ReadVersion();
-            if (version == 0)
-            {
-                return false; // node locked or obsolete — restart
-            }
-
-            var index = node.Find(args.Key, args.Comparer, ref accessor);
-            if (index < 0)
-            {
-                index = ~index - 1;
-            }
-
-            var child = node.GetChild(index, ref accessor);
-            int parentCount = node.GetCount(ref accessor);
-
-            // Validate: node wasn't modified during our unlocked read
-            if (!latch.ValidateVersion(version))
-            {
-                return false; // node modified between version read and data read — restart
-            }
-
-            NodeRelatives.Create(child, index, node, parentCount, ref relatives, out var childRelatives, ref accessor);
-
-            ctx.PathNodes[ctx.Depth] = node;
-            ctx.PathChildIndices[ctx.Depth] = index;
-            ctx.PathVersions[ctx.Depth] = version;
-
-            // Store after Create so lazy-resolved siblings are cached in the stored copy
-            ctx.PathRelatives[ctx.Depth] = relatives;
-
-            node = child;
-            relatives = childRelatives;
-            ctx.Depth++;
-        }
-
-        // Phase 1.5A: Lock leaf with version validation.
-        // Between Phase 1 descent and lock acquisition, a concurrent writer may have split/modified
-        // this leaf. Snapshot the version before locking, then validate after.
-        var leafLatch = node.GetLatch(ref accessor);
-        int leafVersion = leafLatch.ReadVersion();
-        if (leafVersion == 0)
-        {
-            // Leaf is locked or obsolete. SpinWriteLock to wait, then restart.
-            SpinWriteLock(leafLatch);
-            leafLatch.AbortWriteLock();
-            return false;
-        }
-        SpinWriteLock(leafLatch);
-        if (!leafLatch.ValidateVersionLocked(leafVersion))
-        {
-            leafLatch.AbortWriteLock();
-            return false; // restart — leaf was modified between descent and lock
-        }
-
-        // Check if key exists in this leaf
-        var keyIndex = node.Find(args.Key, args.Comparer, ref accessor);
-        if (keyIndex < 0)
-        {
-            node.GetLatch(ref accessor).AbortWriteLock(); // key not found — didn't modify leaf
-            completed = true;
-            return false; // key not found — no merge
-        }
-
-        // Fast path: leaf won't underflow after remove (count > capacity/2) or root leaf (depth == 0).
-        // RemoveLeaf only modifies the leaf in this case (no borrow/merge needed).
-        int count = node.GetCount(ref accessor);
-        if (count > node.GetCapacity() / 2 || ctx.Depth == 0)
-        {
-            bool fastMerged = node.RemoveLeaf(ref args, ref relatives, ref accessor);
-            node.GetLatch(ref accessor).WriteUnlock();
-            completed = true;
-            return fastMerged;
-        }
-
-        // Slow path: leaf may underflow → need neighbors + path for borrow/merge.
-        // Lock leaf neighbors for potential borrow or merge.
-        // AbortWriteLock on failure: no nodes modified yet — avoid spurious version bumps.
-        var leafPrev = node.GetPrevious(ref accessor);
-        var leafNext = node.GetNext(ref accessor);
-        if (leafPrev.IsValid && !leafPrev.GetLatch(ref accessor).TryWriteLock())
-        {
-            node.GetLatch(ref accessor).AbortWriteLock();
-            return false; // restart
-        }
-        if (leafNext.IsValid && !leafNext.GetLatch(ref accessor).TryWriteLock())
-        {
-            if (leafPrev.IsValid)
-            {
-                leafPrev.GetLatch(ref accessor).AbortWriteLock();
-            }
-            node.GetLatch(ref accessor).AbortWriteLock();
-            return false; // restart
-        }
-
-        // Lock path nodes bottom-up with version validation.
-        // Required for ancestor key updates during borrow and merge propagation.
-        // AbortWriteLock on failure: no nodes modified yet — avoid spurious version bumps.
-        for (int i = ctx.Depth - 1; i >= 0; i--)
-        {
-            var pathLatch = ctx.PathNodes[i].GetLatch(ref accessor);
-            if (!pathLatch.TryWriteLock())
-            {
-                for (int j = i + 1; j < ctx.Depth; j++)
-                {
-                    ctx.PathNodes[j].GetLatch(ref accessor).AbortWriteLock();
-                }
-                if (leafNext.IsValid)
-                {
-                    leafNext.GetLatch(ref accessor).AbortWriteLock();
-                }
-                if (leafPrev.IsValid)
-                {
-                    leafPrev.GetLatch(ref accessor).AbortWriteLock();
-                }
-                node.GetLatch(ref accessor).AbortWriteLock();
-                return false; // restart
-            }
-            if (!pathLatch.ValidateVersionLocked(ctx.PathVersions[i]))
-            {
-                pathLatch.AbortWriteLock();
-                for (int j = i + 1; j < ctx.Depth; j++)
-                {
-                    ctx.PathNodes[j].GetLatch(ref accessor).AbortWriteLock();
-                }
-                if (leafNext.IsValid)
-                {
-                    leafNext.GetLatch(ref accessor).AbortWriteLock();
-                }
-                if (leafPrev.IsValid)
-                {
-                    leafPrev.GetLatch(ref accessor).AbortWriteLock();
-                }
-                node.GetLatch(ref accessor).AbortWriteLock();
-                return false; // restart
-            }
-        }
-
-        // All needed nodes locked — Phase 2: Remove at leaf (may borrow/merge)
-        var merged = node.RemoveLeaf(ref args, ref relatives, ref accessor);
-
-        // Phase 2.5: Mark obsolete merged leaf + unlock leaf neighbors + leaf (version bumped by WriteUnlock)
-        var retireEpoch = _segment.Manager.EpochManager.GlobalEpoch;
-        if (merged)
-        {
-            Interlocked.Increment(ref _mergeCount);
-            if (relatives.HasTrueLeftSibling)
-            {
-                // Current node was merged into its left sibling — mark current obsolete
-                node.GetLatch(ref accessor).MarkObsolete();
-                DeferredAdd(node.ChunkId, retireEpoch);
-            }
-            else if (relatives.HasTrueRightSibling && leafNext.IsValid)
-            {
-                // Right sibling was merged into current — mark right sibling obsolete
-                leafNext.GetLatch(ref accessor).MarkObsolete();
-                DeferredAdd(leafNext.ChunkId, retireEpoch);
-            }
-        }
-        if (leafNext.IsValid)
-        {
-            leafNext.GetLatch(ref accessor).WriteUnlock();
-        }
-        if (leafPrev.IsValid)
-        {
-            leafPrev.GetLatch(ref accessor).WriteUnlock();
-        }
-        node.GetLatch(ref accessor).WriteUnlock();
-
-        // Phase 3: Propagate merges upward through internal nodes
-        while (ctx.Depth > 0 && merged)
-        {
-            ctx.Depth--;
-            node = ctx.PathNodes[ctx.Depth];
-            relatives = ctx.PathRelatives[ctx.Depth];
-
-            // Lock siblings that HandleChildMerge might borrow from or merge with
-            NodeWrapper leftSib = relatives.GetLeftSibling(ref accessor);
-            NodeWrapper rightSib = relatives.GetRightSibling(ref accessor);
-            if (leftSib.IsValid)
-            {
-                SpinWriteLock(leftSib.GetLatch(ref accessor));
-            }
-            if (rightSib.IsValid)
-            {
-                SpinWriteLock(rightSib.GetLatch(ref accessor));
-            }
-
-            merged = node.HandleChildMerge(ctx.PathChildIndices[ctx.Depth], ref relatives, ref accessor);
-
-            // Mark obsolete internal node that was merged
-            if (merged)
-            {
-                Interlocked.Increment(ref _mergeCount);
-                if (relatives.HasTrueLeftSibling)
-                {
-                    // Current internal node merged into left sibling
-                    node.GetLatch(ref accessor).MarkObsolete();
-                    DeferredAdd(node.ChunkId, retireEpoch);
-                }
-                else if (relatives.HasTrueRightSibling && rightSib.IsValid)
-                {
-                    // Right sibling merged into current
-                    rightSib.GetLatch(ref accessor).MarkObsolete();
-                    DeferredAdd(rightSib.ChunkId, retireEpoch);
-                }
-            }
-
-            // Unlock siblings + this path node
-            if (rightSib.IsValid)
-            {
-                rightSib.GetLatch(ref accessor).WriteUnlock();
-            }
-            if (leftSib.IsValid)
-            {
-                leftSib.GetLatch(ref accessor).WriteUnlock();
-            }
-            node.GetLatch(ref accessor).WriteUnlock();
-        }
-
-        // Phase 3.5: Unlock remaining path nodes above propagation level
-        while (ctx.Depth > 0)
-        {
-            ctx.Depth--;
-            ctx.PathNodes[ctx.Depth].GetLatch(ref accessor).WriteUnlock();
-        }
-
-        completed = true;
-        return merged;
     }
 
     private NodeWrapper FindLeaf(TKey key, out int index, ref ChunkAccessor accessor)
@@ -2891,559 +1527,6 @@ public abstract partial class BTree<TKey> : IBTree where TKey : unmanaged
         }
 
         return (node.ChunkId, version, keyIndex);
-    }
-
-    /// <summary>
-    /// Compound move for unique indexes: atomically removes the entry at <paramref name="oldKey"/> and inserts it under <paramref name="newKey"/>.
-    /// Uses OLC: same-leaf fast path (single lock), different-leaf (dual lock by ChunkId order).
-    /// Falls back to pessimistic after <see cref="MaxOptimisticRestarts"/>.
-    /// </summary>
-    /// <returns>True if the old key was found and moved; false if old key not found.</returns>
-    public bool Move(TKey oldKey, TKey newKey, int value, ref ChunkAccessor accessor)
-    {
-        // Per-operation accessor for thread safety under OLC (thread-local warm cache)
-        ref var opAccessor = ref _segment.RentWarmAccessor(accessor.ChangeSet);
-        try
-        {
-            for (int attempt = 0; attempt < MaxOptimisticRestarts; attempt++)
-            {
-                // Phase 1: Optimistic descent for both keys
-                var (oldLeafId, oldVersion, oldKeyIndex) = OptimisticDescendToLeaf(oldKey, ref opAccessor);
-                if (oldLeafId == 0)
-                {
-                    if (IsEmpty())
-                    {
-                        return false;
-                    }
-                    Interlocked.Increment(ref _optimisticRestarts);
-                    continue;
-                }
-
-                if (oldKeyIndex < 0)
-                {
-                    // oldKey not found — validate this is not a stale read
-                    var checkLeaf = _storage.LoadNode(oldLeafId);
-                    if (checkLeaf.GetLatch(ref opAccessor).ValidateVersion(oldVersion))
-                    {
-                        return false; // genuinely not found
-                    }
-                    Interlocked.Increment(ref _optimisticRestarts);
-                    continue; // stale read, restart
-                }
-
-                var (newLeafId, newVersion, newKeyIndex) = OptimisticDescendToLeaf(newKey, ref opAccessor);
-                if (newLeafId == 0)
-                {
-                    Interlocked.Increment(ref _optimisticRestarts);
-                    continue; // restart
-                }
-
-                // Phase 2: Lock and mutate
-                if (oldLeafId == newLeafId)
-                {
-                    // Same-leaf fast path: single WriteLock, net count unchanged
-                    var leaf = _storage.LoadNode(oldLeafId);
-                    var latch = leaf.GetLatch(ref opAccessor);
-                    if (!latch.TryWriteLock())
-                    {
-                        Interlocked.Increment(ref _optimisticRestarts);
-                        continue; // contended, restart
-                    }
-
-                    // Validate version (detects concurrent modification between our read and lock)
-                    if (!latch.ValidateVersion(oldVersion | 1))
-                    {
-                        latch.WriteUnlock();
-                        Interlocked.Increment(ref _optimisticRestarts);
-                        continue;
-                    }
-
-                    // Re-find under lock (indices may have shifted)
-                    var oi = leaf.Find(oldKey, Comparer, ref opAccessor);
-                    if (oi < 0)
-                    {
-                        latch.WriteUnlock();
-                        return false; // old key gone
-                    }
-
-                    // Check newKey doesn't already exist BEFORE modifying anything
-                    var ni = leaf.Find(newKey, Comparer, ref opAccessor);
-                    if (ni >= 0)
-                    {
-                        latch.WriteUnlock();
-                        return false; // newKey already exists — no modification
-                    }
-
-                    // Remove old entry and insert new entry
-                    leaf.RemoveAtInternal(oi, ref opAccessor);
-                    // Re-find insertion point after removal (indices shifted)
-                    ni = leaf.Find(newKey, Comparer, ref opAccessor);
-                    ni = ~ni;
-                    leaf.Insert(ni, new KeyValueItem(newKey, value), ref opAccessor);
-
-                    latch.WriteUnlock();
-                    return true;
-                }
-                else
-                {
-                    // Different-leaf path: lock in ChunkId order to prevent deadlocks
-                    var firstId = Math.Min(oldLeafId, newLeafId);
-                    var secondId = Math.Max(oldLeafId, newLeafId);
-                    var firstVersion = oldLeafId == firstId ? oldVersion : newVersion;
-                    var secondVersion = oldLeafId == firstId ? newVersion : oldVersion;
-
-                    var firstLeaf = _storage.LoadNode(firstId);
-                    var secondLeaf = _storage.LoadNode(secondId);
-
-                    var firstLatch = firstLeaf.GetLatch(ref opAccessor);
-                    if (!firstLatch.TryWriteLock())
-                    {
-                        Interlocked.Increment(ref _optimisticRestarts);
-                        continue;
-                    }
-
-                    var secondLatch = secondLeaf.GetLatch(ref opAccessor);
-                    if (!secondLatch.TryWriteLock())
-                    {
-                        firstLatch.WriteUnlock();
-                        Interlocked.Increment(ref _optimisticRestarts);
-                        continue;
-                    }
-
-                    // Validate both versions
-                    if (!firstLatch.ValidateVersion(firstVersion | 1) || !secondLatch.ValidateVersion(secondVersion | 1))
-                    {
-                        secondLatch.WriteUnlock();
-                        firstLatch.WriteUnlock();
-                        Interlocked.Increment(ref _optimisticRestarts);
-                        continue;
-                    }
-
-                    // Identify which is old and which is new
-                    var oldLeaf = oldLeafId == firstId ? firstLeaf : secondLeaf;
-                    var newLeaf = oldLeafId == firstId ? secondLeaf : firstLeaf;
-
-                    // Safety check: if newLeaf is full (insert would overflow) or oldLeaf would underflow, bail to pessimistic which handles structural
-                    // modifications properly
-                    if (newLeaf.GetIsFull(ref opAccessor) || !oldLeaf.GetIsHalfFull(ref opAccessor))
-                    {
-                        secondLatch.WriteUnlock();
-                        firstLatch.WriteUnlock();
-                        break; // fall to pessimistic
-                    }
-
-                    // Re-find under locks
-                    var oi = oldLeaf.Find(oldKey, Comparer, ref opAccessor);
-                    if (oi < 0)
-                    {
-                        secondLatch.WriteUnlock();
-                        firstLatch.WriteUnlock();
-                        return false;
-                    }
-
-                    var ni = newLeaf.Find(newKey, Comparer, ref opAccessor);
-                    if (ni >= 0)
-                    {
-                        // newKey already exists — fail without modification
-                        secondLatch.WriteUnlock();
-                        firstLatch.WriteUnlock();
-                        return false;
-                    }
-                    ni = ~ni;
-
-                    // Remove from old, insert into new
-                    oldLeaf.RemoveAtInternal(oi, ref opAccessor);
-                    newLeaf.Insert(ni, new KeyValueItem(newKey, value), ref opAccessor);
-
-                    secondLatch.WriteUnlock();
-                    firstLatch.WriteUnlock();
-                    return true;
-                }
-            }
-
-            // Pessimistic fallback: full exclusive lock
-            Interlocked.Increment(ref _pessimisticFallbacks);
-            return MovePessimistic(oldKey, newKey, value, ref opAccessor);
-        }
-        finally
-        {
-            _segment.ReturnWarmAccessor();
-        }
-    }
-
-    /// <summary>
-    /// Pessimistic fallback for Move: traverses, removes oldKey, inserts newKey.
-    /// No global lock — concurrency is handled by per-node OLC latches in Remove/Insert.
-    /// </summary>
-    private bool MovePessimistic(TKey oldKey, TKey newKey, int value, ref ChunkAccessor accessor)
-    {
-        try
-        {
-            var oldLeaf = FindLeaf(oldKey, out var oldIndex, ref accessor);
-            if (!oldLeaf.IsValid || oldIndex < 0)
-            {
-                return false;
-            }
-
-            // Check that newKey doesn't already exist
-            var newLeaf = FindLeaf(newKey, out var newIndex, ref accessor);
-            if (newLeaf.IsValid && newIndex >= 0)
-            {
-                return false; // newKey already exists
-            }
-
-            // Remove old entry — use RemoveArguments/RemoveCore for proper structural handling
-            var removeArgs = new RemoveArguments(oldKey, Comparer, ref accessor);
-            RemoveCorePessimistic(ref removeArgs);
-            if (!removeArgs.Removed)
-            {
-                return false;
-            }
-
-            // Insert new entry
-            var insertArgs = new InsertArguments(newKey, value, Comparer, ref accessor);
-            AddOrUpdateCorePessimistic(ref insertArgs);
-            SyncHeader(ref accessor);
-            return true;
-        }
-        finally
-        {
-            DeferredReclaim();
-        }
-    }
-
-    /// <summary>
-    /// Compound move for AllowMultiple indexes: removes <paramref name="elementId"/>/<paramref name="value"/> from <paramref name="oldKey"/>'s buffer and
-    /// appends <paramref name="value"/> under <paramref name="newKey"/>.
-    /// Returns the new element ID and both HEAD buffer IDs for inline TAIL tracking.
-    /// </summary>
-    public int MoveValue(TKey oldKey, TKey newKey, int elementId, int value,
-        ref ChunkAccessor accessor, out int oldHeadBufferId, out int newHeadBufferId, bool preserveEmptyBuffer = false)
-    {
-        // Per-operation accessor for thread safety under OLC (thread-local warm cache)
-        ref var opAccessor = ref _segment.RentWarmAccessor(accessor.ChangeSet);
-        try
-        {
-            for (int attempt = 0; attempt < MaxOptimisticRestarts; attempt++)
-            {
-                // Phase 1: Optimistic descent for both keys
-                var (oldLeafId, oldVersion, oldKeyIndex) = OptimisticDescendToLeaf(oldKey, ref opAccessor);
-                if (oldLeafId == 0)
-                {
-                    Interlocked.Increment(ref _optimisticRestarts);
-                    continue;
-                }
-
-                if (oldKeyIndex < 0)
-                {
-                    var checkLeaf = _storage.LoadNode(oldLeafId);
-                    if (checkLeaf.GetLatch(ref opAccessor).ValidateVersion(oldVersion))
-                    {
-                        oldHeadBufferId = -1;
-                        newHeadBufferId = -1;
-                        return -1; // old key genuinely not found
-                    }
-                    Interlocked.Increment(ref _optimisticRestarts);
-                    continue;
-                }
-
-                var (newLeafId, newVersion, newKeyIndex) = OptimisticDescendToLeaf(newKey, ref opAccessor);
-                if (newLeafId == 0)
-                {
-                    Interlocked.Increment(ref _optimisticRestarts);
-                    continue;
-                }
-
-                // Phase 2: Lock and mutate
-                if (oldLeafId == newLeafId)
-                {
-                    var leaf = _storage.LoadNode(oldLeafId);
-                    var latch = leaf.GetLatch(ref opAccessor);
-                    if (!latch.TryWriteLock())
-                    {
-                        Interlocked.Increment(ref _optimisticRestarts);
-                        continue;
-                    }
-
-                    if (!latch.ValidateVersion(oldVersion | 1))
-                    {
-                        latch.WriteUnlock();
-                        Interlocked.Increment(ref _optimisticRestarts);
-                        continue;
-                    }
-
-                    // Re-find oldKey under lock
-                    var oi = leaf.Find(oldKey, Comparer, ref opAccessor);
-                    if (oi < 0)
-                    {
-                        latch.WriteUnlock();
-                        oldHeadBufferId = -1;
-                        newHeadBufferId = -1;
-                        return -1;
-                    }
-
-                    // Remove element from old buffer
-                    var oldBufferId = leaf.GetItem(oi, ref opAccessor).Value;
-                    var res = _storage.RemoveFromBuffer(oldBufferId, elementId, value, ref opAccessor);
-                    oldHeadBufferId = oldBufferId;
-
-                    if (res == -1)
-                    {
-                        latch.WriteUnlock();
-                        newHeadBufferId = -1;
-                        return -1; // element not found in buffer
-                    }
-
-                    // Find or prepare newKey
-                    var ni = leaf.Find(newKey, Comparer, ref opAccessor);
-                    int newBufferId;
-                    int newElementId;
-                    if (ni >= 0)
-                    {
-                        // newKey exists — append to its buffer
-                        newBufferId = leaf.GetItem(ni, ref opAccessor).Value;
-                        newElementId = _storage.Append(newBufferId, value, ref opAccessor);
-                    }
-                    else
-                    {
-                        // newKey doesn't exist — need to insert a new key entry
-                        // If leaf is full and we won't reclaim a slot, bail to pessimistic.
-                        // We can only reclaim when res==0 (old buffer empty) AND !preserveEmptyBuffer.
-                        if (leaf.GetIsFull(ref opAccessor) && (res != 0 || preserveEmptyBuffer))
-                        {
-                            // Undo the buffer removal — re-add the element
-                            _storage.Append(oldBufferId, value, ref opAccessor);
-                            latch.WriteUnlock();
-                            break; // fall to pessimistic
-                        }
-
-                        newBufferId = _storage.CreateBuffer(ref opAccessor);
-                        newElementId = _storage.Append(newBufferId, value, ref opAccessor);
-                        ni = ~ni;
-                        // If old buffer empty (res==0) and not preserving, remove old key first to free a slot
-                        if (res == 0 && !preserveEmptyBuffer)
-                        {
-                            oi = leaf.Find(oldKey, Comparer, ref opAccessor);
-                            if (oi >= 0)
-                            {
-                                leaf.RemoveAtInternal(oi, ref opAccessor);
-                                _storage.DeleteBuffer(oldBufferId, ref opAccessor);
-                                Interlocked.Decrement(ref _count);
-                            }
-                            // Re-find insertion point after removal
-                            ni = leaf.Find(newKey, Comparer, ref opAccessor);
-                            ni = ~ni;
-                            res = -2; // sentinel: old key already cleaned up
-                        }
-                        leaf.Insert(ni, new KeyValueItem(newKey, newBufferId), ref opAccessor);
-                        Interlocked.Increment(ref _count);
-                    }
-                    newHeadBufferId = newBufferId;
-
-                    // If old buffer is now empty and not yet cleaned up, remove the BTree entry for oldKey
-                    if (res == 0 && !preserveEmptyBuffer)
-                    {
-                        // Re-find oldKey (index may have shifted after insert)
-                        oi = leaf.Find(oldKey, Comparer, ref opAccessor);
-                        if (oi >= 0)
-                        {
-                            leaf.RemoveAtInternal(oi, ref opAccessor);
-                            _storage.DeleteBuffer(oldBufferId, ref opAccessor);
-                            Interlocked.Decrement(ref _count);
-                        }
-                    }
-
-                    latch.WriteUnlock();
-                    SyncHeader(ref opAccessor);
-                    return newElementId;
-                }
-                else
-                {
-                    // Different-leaf path: lock in ChunkId order
-                    var firstId = Math.Min(oldLeafId, newLeafId);
-                    var secondId = Math.Max(oldLeafId, newLeafId);
-                    var firstVersion = oldLeafId == firstId ? oldVersion : newVersion;
-                    var secondVersion = oldLeafId == firstId ? newVersion : oldVersion;
-
-                    var firstLeaf = _storage.LoadNode(firstId);
-                    var secondLeaf = _storage.LoadNode(secondId);
-
-                    var firstLatch = firstLeaf.GetLatch(ref opAccessor);
-                    if (!firstLatch.TryWriteLock())
-                    {
-                        Interlocked.Increment(ref _optimisticRestarts);
-                        continue;
-                    }
-
-                    var secondLatch = secondLeaf.GetLatch(ref opAccessor);
-                    if (!secondLatch.TryWriteLock())
-                    {
-                        firstLatch.WriteUnlock();
-                        Interlocked.Increment(ref _optimisticRestarts);
-                        continue;
-                    }
-
-                    if (!firstLatch.ValidateVersion(firstVersion | 1) || !secondLatch.ValidateVersion(secondVersion | 1))
-                    {
-                        secondLatch.WriteUnlock();
-                        firstLatch.WriteUnlock();
-                        Interlocked.Increment(ref _optimisticRestarts);
-                        continue;
-                    }
-
-                    var oldLeaf = oldLeafId == firstId ? firstLeaf : secondLeaf;
-                    var newLeaf = oldLeafId == firstId ? secondLeaf : firstLeaf;
-
-                    // Pre-check: if newLeaf is full and newKey doesn't exist, bail to pessimistic (we'd need to insert a new entry which could cause overflow)
-                    if (newLeaf.GetIsFull(ref opAccessor))
-                    {
-                        var preNi = newLeaf.Find(newKey, Comparer, ref opAccessor);
-                        if (preNi < 0)
-                        {
-                            secondLatch.WriteUnlock();
-                            firstLatch.WriteUnlock();
-                            break; // fall to pessimistic
-                        }
-                    }
-
-                    // Remove element from old buffer
-                    var oi = oldLeaf.Find(oldKey, Comparer, ref opAccessor);
-                    if (oi < 0)
-                    {
-                        secondLatch.WriteUnlock();
-                        firstLatch.WriteUnlock();
-                        oldHeadBufferId = -1;
-                        newHeadBufferId = -1;
-                        return -1;
-                    }
-
-                    var oldBufferId = oldLeaf.GetItem(oi, ref opAccessor).Value;
-                    var res = _storage.RemoveFromBuffer(oldBufferId, elementId, value, ref opAccessor);
-                    oldHeadBufferId = oldBufferId;
-
-                    if (res == -1)
-                    {
-                        secondLatch.WriteUnlock();
-                        firstLatch.WriteUnlock();
-                        newHeadBufferId = -1;
-                        return -1;
-                    }
-
-                    // Append to new buffer
-                    var ni = newLeaf.Find(newKey, Comparer, ref opAccessor);
-                    int newBufferId;
-                    int newElementId;
-                    if (ni >= 0)
-                    {
-                        newBufferId = newLeaf.GetItem(ni, ref opAccessor).Value;
-                        newElementId = _storage.Append(newBufferId, value, ref opAccessor);
-                    }
-                    else
-                    {
-                        newBufferId = _storage.CreateBuffer(ref opAccessor);
-                        newElementId = _storage.Append(newBufferId, value, ref opAccessor);
-                        ni = ~ni;
-                        newLeaf.Insert(ni, new KeyValueItem(newKey, newBufferId), ref opAccessor);
-                        Interlocked.Increment(ref _count);
-                    }
-                    newHeadBufferId = newBufferId;
-
-                    // If old buffer is now empty, remove the BTree entry
-                    if (res == 0 && !preserveEmptyBuffer)
-                    {
-                        oi = oldLeaf.Find(oldKey, Comparer, ref opAccessor);
-                        if (oi >= 0)
-                        {
-                            oldLeaf.RemoveAtInternal(oi, ref opAccessor);
-                            _storage.DeleteBuffer(oldBufferId, ref opAccessor);
-                            Interlocked.Decrement(ref _count);
-                        }
-                    }
-
-                    secondLatch.WriteUnlock();
-                    firstLatch.WriteUnlock();
-                    SyncHeader(ref opAccessor);
-                    return newElementId;
-                }
-            }
-
-            // Pessimistic fallback
-            Interlocked.Increment(ref _pessimisticFallbacks);
-            return MoveValuePessimistic(oldKey, newKey, elementId, value, ref opAccessor, out oldHeadBufferId, out newHeadBufferId, preserveEmptyBuffer);
-        }
-        finally
-        {
-            _segment.ReturnWarmAccessor();
-        }
-    }
-
-    /// <summary>
-    /// Pessimistic fallback for MoveValue: removes element from old buffer,
-    /// appends to new buffer, handles empty-buffer cleanup.
-    /// No global lock — concurrency is handled by per-node OLC latches in Remove/Insert.
-    /// </summary>
-    private int MoveValuePessimistic(TKey oldKey, TKey newKey, int elementId, int value, ref ChunkAccessor accessor, out int oldHeadBufferId, 
-        out int newHeadBufferId, bool preserveEmptyBuffer = false)
-    {
-        try
-        {
-            var oldLeaf = FindLeaf(oldKey, out var oldIndex, ref accessor);
-            if (!oldLeaf.IsValid || oldIndex < 0)
-            {
-                oldHeadBufferId = -1;
-                newHeadBufferId = -1;
-                return -1;
-            }
-
-            // Remove element from old buffer
-            var oldBufferId = oldLeaf.GetItem(oldIndex, ref accessor).Value;
-            var res = _storage.RemoveFromBuffer(oldBufferId, elementId, value, ref accessor);
-            oldHeadBufferId = oldBufferId;
-
-            if (res == -1)
-            {
-                newHeadBufferId = -1;
-                return -1;
-            }
-
-            // Append to new key's buffer
-            var newLeaf = FindLeaf(newKey, out var newIndex, ref accessor);
-            int newBufferId;
-            int newElementId;
-            if (newLeaf.IsValid && newIndex >= 0)
-            {
-                // newKey exists — append to its buffer
-                newBufferId = newLeaf.GetItem(newIndex, ref accessor).Value;
-                newElementId = _storage.Append(newBufferId, value, ref accessor);
-            }
-            else
-            {
-                // newKey doesn't exist — create buffer and insert via AddOrUpdateCore
-                newBufferId = _storage.CreateBuffer(ref accessor);
-                newElementId = _storage.Append(newBufferId, value, ref accessor);
-                var insertArgs = new InsertArguments(newKey, newBufferId, Comparer, ref accessor);
-                AddOrUpdateCorePessimistic(ref insertArgs);
-            }
-            newHeadBufferId = newBufferId;
-
-            // If old buffer is now empty, remove the BTree entry
-            if (res == 0 && !preserveEmptyBuffer)
-            {
-                var removeArgs = new RemoveArguments(oldKey, Comparer, ref accessor);
-                RemoveCorePessimistic(ref removeArgs);
-                if (removeArgs.Removed)
-                {
-                    _storage.DeleteBuffer(oldBufferId, ref accessor);
-                }
-            }
-
-            SyncHeader(ref accessor);
-            return newElementId;
-        }
-        finally
-        {
-            DeferredReclaim();
-        }
     }
 
     #endregion
