+++ src/Typhon.Engine/Data/Transaction/ComponentInfo.cs
@@ -3,7 +3,6 @@
 using System;
 using System.Collections.Generic;
 using System.Runtime.InteropServices;
-using Typhon.Engine.BPTree;
 
 namespace Typhon.Engine;
