+++ test/Typhon.Benchmark/BTreeDeleteProfile.cs
@@ -4,7 +4,6 @@ using System;
 using System.Diagnostics;
 using System.IO;
 using Typhon.Engine;
-using Typhon.Engine.BPTree;
 
 namespace Typhon.Benchmark;
