+++ test/Typhon.Benchmark/BTreeEnumerationBenchmarks.cs
@@ -3,7 +3,6 @@ using BenchmarkDotNet.Attributes;
 using System.Threading;
 using System.Threading.Tasks;
 using Typhon.Engine;
-using Typhon.Engine.BPTree;
 
 namespace Typhon.Benchmark;
