+++ test/Typhon.Engine.Tests/Data/BTreeTests.cs
@@ -6,7 +6,6 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using Typhon.Engine.BPTree;
 
 namespace Typhon.Engine.Tests;
