+++ test/Typhon.Engine.Tests/Data/OlcBTreeStressTests.cs
@@ -3,7 +3,6 @@ using NUnit.Framework;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
-using Typhon.Engine.BPTree;
 
 namespace Typhon.Engine.Tests;
