+++ test/Typhon.Engine.Tests/Data/OlcBTreeTests.cs
@@ -4,7 +4,6 @@ using System;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Typhon.Engine.BPTree;
 
 namespace Typhon.Engine.Tests;
