+++ test/Typhon.Engine.Tests/Data/OlcLatchTests.cs
@@ -4,7 +4,6 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Threading;
 using System.Threading.Tasks;
-using Typhon.Engine.BPTree;
 
 namespace Typhon.Engine.Tests;
